add erd png file
