+++ app/controllers/application_controller.rb
@@ -1,19 +1,12 @@
-module ApplicationHelper
-    def menu_link_to(link_text, link_path)
-      class_name = current_page?(link_path) ? 'menu-item active' : 'menu-item'
-  
-      content_tag(:div, class: class_name) do
-        link_to link_text, link_path
-      end
-    end
-  
-    def like_or_dislike_btn(post)
-      like = Like.find_by(post: post, user: current_user)
-      if like
-        link_to('Dislike!', post_like_path(id: like.id, post_id: post.id), method: :delete)
-      else
-        link_to('Like!', post_likes_path(post_id: post.id), method: :post)
-      end
-    end
+class ApplicationController < ActionController::Base
+  protect_from_forgery with: :exception
+
+  before_action :configure_permitted_parameters, if: :devise_controller?
+
+  protected
+
+  def configure_permitted_parameters
+    devise_parameter_sanitizer.permit(:sign_up) { |u| u.permit(:name, :email, :password, :password_confirmation) }
+    devise_parameter_sanitizer.permit(:account_update) { |u| u.permit(:name, :email, :password, :current_password) }
   end
-  
+end
